Moved PBWT sort functions and removed unused class
